+++ src/ec4py/lsv_data.py
@@ -5,6 +5,7 @@
 from __future__ import annotations
 import math
 import numpy as np
+import pandas as pd
 # from scipy import integrate
 from scipy.signal import savgol_filter 
 
@@ -16,7 +17,7 @@ from .util_voltammetry import Voltammetry, OFFSET_AT_E_MIN, OFFSET_AT_E_MAX, OFF
 from .ec_setup import EC_Setup
 from .util import extract_value_unit     
 from .util import Quantity_Value_Unit as QV
-from .util_graph import plot_options,quantity_plot_fix, make_plot_2x,make_plot_1x
+from .util_graph import plot_options,quantity_plot_fix, make_plot_2x,make_plot_1x,should_plot_be_made
 from .analysis_tafel import Tafel
 from .analysis_levich import diffusion_limit_corr
 
@@ -284,18 +285,6 @@ class LSV_Data(Voltammetry):
                 self.i = v
         return 
         
-        #norm_factor = self.get_norm_factor(norm_to)
-        #print(norm_factor)
-        #if norm_factor:
-        #    self.i = self.i / float(norm_factor)
-             
-        #norm_factor_inv = norm_factor ** -1
-        #    current = QV(1,self.i_unit, self.i_label) / norm_factor
-         
-        #    self.i_label = current.quantity
-        #    self.i_unit = current.unit
-        
-        #return 
     
     ############################################################################        
     def plot(self,*args, **kwargs):
@@ -310,24 +299,28 @@ class LSV_Data(Voltammetry):
             line, ax: line and ax handlers
         
         '''
-        data = copy.deepcopy(self)
-        options = plot_options(kwargs)
-        options.set_title(self.setup_data.name)
-        options.name = self.setup_data.name
-        options.legend = self.legend(*args, **kwargs)
-        #if options.legend == "_" :
-        #        data_plot_kwargs["legend"] = data.setup_data.name
-        #data
-        data.norm(args)
-        data.set_active_RE(args)
-        options.x_data = data.E
-        options.y_data = data.i
-                
-        options.set_x_txt(data.E_label, data.E_unit)
-        options.set_y_txt(data.i_label, data.i_unit) 
-        
-            
-        return options.exe()
+        if should_plot_be_made(*args):
+            data = copy.deepcopy(self)
+            options = plot_options(kwargs)
+            if self.is_MWE:
+                options.set_title(f"{self.setup_data.name}#{self.setup_data._MWE_CH}")
+            else:
+                options.set_title(self.setup_data.name)
+            #options.name = self.setup_data.name
+            options.legend = self.legend(*args, **kwargs)
+            #if options.legend == "_" :
+            #        data_plot_kwargs["legend"] = data.setup_data.name
+            #data
+            data.norm(args)
+            data.set_active_RE(args)
+            options.x_data = data.E
+            options.y_data = data.i
+                    
+            options.set_x_txt(data.E_label, data.E_unit)
+            options.set_y_txt(data.i_label, data.i_unit) 
+            return options.exe()
+        else:
+            return None,None
     
     
     def set_active_RE(self,shift_to:str|tuple = None):
@@ -358,7 +351,7 @@ class LSV_Data(Voltammetry):
         Args:
             E (float): potential where to get the current. 
         Returns:
-            _type_: _description_
+             Quantity_Value_Unit: The current, units and label.
         """
         lsv = copy.deepcopy(self)
         lsv.norm(args)
@@ -372,6 +365,69 @@ class LSV_Data(Voltammetry):
         return QV(lsv.i[index],lsv.i_unit,lsv.i_label)
     ###########################################################################################
 
+    def get_E_of_max_i(self, E1:float,E2:float,*args,**kwargs):
+        """get the potential of maximum current in a range.
+
+        Args:
+            E1 (float): _description_
+            E2 (float): _description_
+
+        Returns:
+             (Quantity_Value_Unit | None): _description_
+        """
+        
+        index1 = self.get_index_of_E(E1)
+        index2 = self.get_index_of_E(E2)
+        index_max=max(index1,index2)
+        index_min=min(index1,index2)
+        #print(index_min,index_max)
+        index_of_max=index_min
+        max_i =-1e100
+        for index in range(index_min,index_max):
+            if max_i<self.i[index] and not np.isnan(self.i[index]) :
+                index_of_max = index
+                max_i =self.i[index]
+        if max_i > -1e100:    
+            index_E = index_of_max
+            #index_E = self.i[index_min:index_max].argmax()
+            #print("index",index_E,"E", self.E[index_E])
+            return QV(self.E[index_E],self.E_unit,self.E_label)
+        else:
+            return None
+    ###########################################################################################
+
+    def get_E_of_min_i(self, E1:float,E2:float,*args,**kwargs):
+        """get the potential of minimum current in a range.
+
+        Args:
+            E1 (float): Start potential
+            E2 (float): End potential
+
+        Returns:
+           (Quantity_Value_Unit | None): The voltage of min, or None, if not found
+        """
+                
+        index1 = self.get_index_of_E(E1)
+        index2 = self.get_index_of_E(E2)
+        index_max=max(index1,index2)
+        index_min=min(index1,index2)
+        #print(index_min,index_max)
+        index_of_min=index_min
+        min_i =1e100
+        for index in range(index_min,index_max):
+            if self.i[index]<min_i and not np.isnan(self.i[index]) :
+                index_of_min = index
+                min_i =self.i[index]
+        if min_i < 1e100:    
+            index_E = index_of_min
+            #index_E = self.i[index_min:index_max].argmax()
+            #print("index",index_E,"E", self.E[index_E])
+            return QV(self.E[index_E],self.E_unit,self.E_label)
+        else:
+            return None
+    
+    ##################################################################################################
+    
     def integrate(self, start_E:float, end_E:float, show_plot: bool = False, *args, **kwargs):
         """Integrate Current between the voltage limit using cumulative_simpson
 
@@ -454,3 +510,19 @@ class LSV_Data(Voltammetry):
         data_plot.legend()
     
         return Tafel_slope
+    
+    
+    def export_DataFrame(self):
+        size = [len(Voltammetry().E),2]
+        m = np.zeros(size)
+        col_names= list("E")
+        #print(m.shape,len(self.datas))
+        m[:,0]=Voltammetry().E
+        
+        
+            #print(x,self.datas[x].i.shape)
+        m[:,1] = self.i
+        col_names.append(f"{self.i_label}_{self.name} / {self.i_unit}")
+        #print(col_names)
+        df = pd.DataFrame.from_records(m,columns=col_names)
+        return df