+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - 
 ### Changed
 
+## [0.4.11] - 
+### Added
+  - Export to Pandas for CV_Datas and LSV_Datas
+  - CV_Data and CV_Datas can now do IR-compensation
+### Changed
+
 ## [0.4.10] - 
 ### Added
   - Steps -